fix: apply userAgentBase change and move defaultHeaders tests to shared/common.
