+++ packages/shared/common/src/api/platform/Info.ts
@@ -45,6 +45,12 @@ export interface SdkData {
   version?: string;
 
   /**
+   * If this is a top-level (not a wrapper) SDK this will be used to create the user agent string.
+   * It will take the form 'userAgentBase/version`.
+   */
+  userAgentBase?: string;
+
+  /**
    * Name of the wrapper SDK if present.
    */
   wrapperName?: string;
